+++ packages/headless/doc-parser/src/function-resolver.ts
@@ -5,8 +5,7 @@ import {
 } from '@microsoft/api-extractor-model';
 import {DocComment} from '@microsoft/tsdoc';
 import {findApi} from './api-finder';
-import {ObjEntity} from './entity';
-import {buildEntity, buildFuncEntity} from './entity-builder';
+import {buildEntity, buildFuncEntity, buildObjEntity} from './entity-builder';
 import {resolveParams} from './function-param-resolver';
 import {resolveInterfaceMembers} from './interface-resolver';
 
@@ -32,7 +31,7 @@ export function resolveFunction(
 function buildObjEntityFromInterface(
   entryPoint: ApiEntryPoint,
   apiInterface: ApiInterface
-): ObjEntity {
+) {
   const name = apiInterface.name;
   const members = resolveInterfaceMembers(entryPoint, apiInterface);
   const entity = buildEntity({
@@ -42,5 +41,5 @@ function buildObjEntityFromInterface(
     comment: (apiInterface.tsdocComment as unknown) as DocComment,
   });
 
-  return {...entity, members};
+  return buildObjEntity({entity, members, typeName: name});
 }